// src/app/core/models/registro/User.ts
export class User {
    id?: number;
    legacy_user_id?: number;
    full_name?: String;
    display_name?: String;
    birth_date?: String;
    age?: number;
    gender?: String;
    shirt_size?: String;
    coming_from?: String;
    allergies?: String;
    guardian_phone?: String;
    church?: String;
    registered_at?: String;
    guardian_name?: String;
    guardian_email?: String;
    emailTutor?: String;
    medications?: String;
    phone?: String;
    user_status?: String;
    accepted_policies?: number | boolean;
    nombre?: String;
    nick?: String;
    year?: number;
    month?: number;
    day?: number;
    fechaNac?: Date;
    edad?: number;
    sexo?: String;
    talla?: String;
    vienesDe?: String;
    alergias?: String;
    tutorNombre?: String;
    tutorTelefono?: String;
    iglesia?: String;
    email?: String;
    whatsapp?: String;
    facebook?: String;
    instagram?: String;
    aceptaPoliticas?: boolean;
    password?: String;
    updatePassword?: boolean;
    medicamentos?: String;
    telefono?: String;
}